+++ tracing/tracer.go
@@ -1,0 +1,8 @@
+package tracing
+
+import "github.com/opentracing/opentracing-go"
+
+// NewTracer returns the registered global tracer.
+func NewTracer() opentracing.Tracer {
+	return opentracing.GlobalTracer()
+}