+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 
-## Unreleased
+## 0.2.0 - 2017-10-12
 
 ### Added